Add nc to the packages to install during upgrade
Upgrade uses nc, but doesn't enforce its presence.
This makes sure nc is installed.

On top of it, the filenames used are ubuntu ones,
so this moves to use a var to find which filename
should be used for centos.

See also the issue here:
http://logs.openstack.org/periodic/periodic-openstack-ansible-
upgrade-ceph-master-centos-7/d2fc540/
console.html#_2017-05-30_08_23_40_475244

Change-Id: I3ee2dfd6e435fb4d92a57e2debf7f001f7c03f98