Set ceph_mons in group_vars
ceph_mons is used by the OSA ceph_client role, since it does not
support consuming the list of mons from a group name which is what
upstream ceph-ansible uses. So we will generate the list of mons
using the group name and prepare it for ceph_client.

This defaults the configuration to use the integrated ceph-ansible
managed ceph cluster deployment in OSA. However, if the OSA deployer
is using an externally managed cluster, they should still be able
to override ceph_mons in user_variables.

Change-Id: Ida44228c2d2919990bac46936317b6093d3ff115