Add support for python 3 inventory tests
Define a new tox environment for inventory the inventory tests with the
Python 3.5 interpreter.

This could not be done with [1] due to the way generative environments
create new environments, rather than modifying already existing and
named ones.

For now, the py3-inventory env is not added to linters, in order to keep
it from failing the gates.

1:
https://tox.readthedocs.io/en/latest/config.html#generating-environments-conditional-settings

Change-Id: Ife1ecc6c36c47f38c00bd19a193a637e39a8b4b8