Revert "Reduce container counts for cinder, ironic and trove"
A fully hyper-converged deployment is fine, but this has not
been done with an agreed spec, nor has it been completed.

This patch is being reverted with the agreement that there
will be a spec and we can tackle this properly in Queens.

This reverts commit b769a6e0c128b96b23dd696a305145f2077c3151.

Change-Id: Ia7ac1084db7610800dca7a860fddc2ad825fa101
(cherry picked from commit 4cba47c1d3a186d189c75e903cbb8be39602fa0f)