Fix race condition regression
The building of the group "repo_masters" relies on going through
the inventory, but we never check if the hosts are added in a
particular order. While it should follow the ansible inventory
order (that's the play order by default), the inclusion of the
host to a group could happen at any time, because the play runs
all the hosts in parallel.

If we ensure the hosts are run serially, the addition of the
hosts to the architecture dependant repo_ groups will always
be done in the right order.

We can then use these groups to find which host is the first one,
and use it as "repo_master".

Change-Id: I51803dab20e85bd6c100821ce5c1c1ad3226ba29