scripts: sources-branch-updater-lib.sh: Avoid using 'dirname $0'
Since this library can be sourced by external scripts, 'dirname $0'
can't be used reliably to determine where the rest of the OSA helper
scripts are located since $0 points to the external script.
It's best to use the BASH_SOURCE array to determine the file in which
the current function is defined.

Change-Id: If48ee9cb472de811076d7d3adec8f3c9b476b833