Implement rolling upgrades for keystone
The current rolling upgrade implementation
relies on the role to orchestrate the rolling
upgrade. When the role is executed using
playbook serialisation, the db sync contract
is executed before all hosts are upgraded,
potentially resulting in data corruption.

This patch implements the use of playbook
serialisation for the upgrade and fact checking
to determine when it is safe to execute the db
sync contract.

Additionally, tasks which only need to be
executed against a single keystone host
in the invevntory are specifically targeted
rather than being skipped on the other hosts.
This should provide a slight performance
improvement.

Depends-On: I5650f16b9a115bd392012b743788057a94d09226
Change-Id: Iac59e792a642a9e57d6d279a7d4b3d41fe419b38
(cherry picked from commit ecf32e20784d733fa89e1b0a392fd4b565ca4f41)