Correctly map nova git sources
For any packages the nova role uses which are
not proprietary, the repo buid process will
need to git clone and build them from source.

In a situation where a nova control plane is
built, but no compute hosts are assigned, the
repo build currently fails due to the current
mapping of these git sources to compute hosts.

With this patch we ensure that the git sources
are cloned and the packages built as long as
there are any nova hosts in the inventory.

Change-Id: Ief9da2c8c555b51a06af9f7cd6145f9d74c14478
Closes-Bug: 1701468
(cherry picked from commit e3524ec0d2fc2cd9ea8934c3e791079b8c24b768)