Remove run-playbooks
The run-playbooks script was required in order to
implement the successerator. This is no longer
required as the implementation is much more
reliable than it was when the tool was introduced.

In order to stop masking the few failures being
masked by the use of the successerator, it is
being removed.

Also, in order to make development testing more
like a production deployment, and the AIO
workflow be a little more like a production
deployment workflow, the run-playbooks script is
being removed.

Change-Id: I961b7a17f180e2e89eb8faac8bed4dfc7cfdf719