Fixes deploy-guide links
Due to changes from the documentation migration process, the
tag :deploy_guide: is no longer functioning appropriately.
This patch fixes the links to point to /draft/ until the issue
with :deploy_guide: can be solved.

This patch also re-links to the Mitaka documentation which was
previously left behind in the server migration.

Change-Id: If7c6af44557945dd8f2fca25b69965b1fb9bf295