Revert "Update setuptools to 36.2.0"
Our pip, setuptools, wheel install process is seperate
from the installation of other packages, so it still
works. The problem comes in when you try to up/downgrade
setuptools at the same time as installing other packages.
This problem was prevalent when doing docs/linters tests.

This partially reverts commit
504da781ccb68481261a523afedf8210bdc5f36c and removes the
constraint from the tox config so that doc and linter
tests will not try to force a up/downgrade for these
packages.

Change-Id: I27f843d443b32c52b5f9ec6be581a9366c8dbcf5