Add python package for ansible bootstrap
In the current ansible bootstrap script, python3 is
exclusively installed resulting in there being no
symlink for /usr/bin/python to the appropriate
python binary. This causes failures for anything
which relies on using the 'python' command (instead
of python3), resulting in failure if the python
packages was not previously installed in some other
way.

Installing the 'python-minimal' package results in
python2 being installed and symlinked, but it does
not cause the Ansible venv to be bootstrapped any
differently - it still uses python3 and the venv
is setup with a symlink from python -> python3.

We therefore include the 'python-minimal' package
in the bootstrap so that anything that needs access
to the python binary can still work. In the future
Ubuntu may change where it symlinks to, but by
then all our tooling should be capable of using
the new python version.

Closes-Bug: #1704546
Change-Id: I13f70040a0a3b4d1c9434078aa6d6888abf45787