Revert "Update setuptools to 36.2.0"
Our pip, setuptools, wheel install process is seperate
from the installation of other packages, so it still
works. The problem comes in when you try to up/downgrade
setuptools at the same time as installing other packages.
This problem was prevalent when doing docs/linters tests.

This reverts commit d4ae08646d1c1192e5806be2f81b1748f520fd39
and removes the constraint from the tox config so that doc
and linter tests will not try to force a up/downgrade for
these packages.

Change-Id: I27f843d443b32c52b5f9ec6be581a9366c8dbcf5