Provide default python path
When doing an upgrade, the master scripts library
is loaded into memory and used when bootstrapping
the previous version's ansible. The get_pip
therefore fails due to not being given the path
to the python to use.

This patch ensures that the get_pip function has
a default to prevent the unbound variable error.

Change-Id: I78ab63a7e693fcd89026a1d726ec3354653accea