Move Upgrade branch change above initial library source
Upgrades fail on master due to changes to the library functions,
specifically "get_pip" which now requires a variable passed to it. Since
the library was sourced before it gets called in the Ocata branch this
caused failures when the function was called.

To fix this we should source the library after we change branches, and
not before. Additionally, we should source the scripts-library.sh from
master once we change back to the master branch.

Change-Id: If3fedfcf2dd5b13c918812fcb9aca7efea7d0f26
(cherry picked from commit ece5067966ca769a892e20b938fdb9d2589b49a4)