group_vars: repo_all: Always build cryptography from source
cryptography may bundle openssl in the wheel and that causes symbol
conflicts if a different openssl is provided by the distribution.
As such, it's probably safer to re-build cryptography ourselves just
to be sure that the correct distro libraries are used. This has been
addressed in openstack-ansible-tests/test-vars.yaml
(https://review.openstack.org/#/c/486580/) to fix the CI tests but the
problem is also present on regular deployments so we set it in the
group_variables for the repo_all group of hosts so it's built from
source in the wheel repository.

Related-Bug: 1705521
Link: https://github.com/pyca/cryptography/issues/3804
Change-Id: I54ba3c1fa48a2f4c633930bc7e8cc65397f86659