Remove trove_regular_user
Remove the unnecessary trove_regular_user.  The documenation that was
referenced [1] when this was added to the role is intended to configure
Trove for development purposes. The trove_regular_user is not used by the
Trove service and is only being created to give the developer a non-admin
user to use for testing.

[1] https://docs.openstack.org/trove/latest/install/manual_install.html

Change-Id: Ic71216c21092a22105ad56ef98e1554dff48f0b0