Use ansible-runtime python for internal scripts
The ansible runtime python needs to be used when running local scripts
so that our libs cared for. Presently when the pw-token-gen script is
run it stacktraces on a new 16.04 host when py3.5 and py2.7 are present
on the same host. The traceback indicates that the Crypto lib is missing
which is true for py2.7 as we now prefer py3.5.

Additionally the pw-token-gen script needs two updates to be py3.5
compatible.

Change-Id: I5535c2076ebb3fa5b2a5d5fc42bdc5095871c7d5
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>