Use LXC reverse proxy in OpenStack-CI
This patch implements the use of the images.linuxcontainers.org
reverse proxy when in nodepool. The same implementation as used
for MariaDB and UCA is used - this is not optimal as nodepool
settings may change, but this will do for now.

Depends-On: Ia6b149f157e5697d36fdea3dc65cb2ac4815a1a4
Change-Id: I558663e31264e35a4916c7b636a09cee3910408b
(cherry picked from commit 2b2bd4de69e1844f4801b0bb61e32c64ff8a942d)