Update the playbook serial settings
Many of our playbooks enfore the serial deployment of services. While
this is useful in certain cases It's super slow on large scale
deployments. This change modifies our serial settings so that many of
the services are deployed against 1 node first and then all the rest of
the services where it makes.

Change-Id: I842e98ebc29e0da0c3d7af3fd1d7f5592dfb3aa3
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>