Vagrantfile: Increase VM memory size to 12G
Running AIO on Vagrant results in the entire 8G of memory being used.
This makes the VM really slow after a while and as a result of which
various services may start to fail due to timeouts. It's best to
increase the VM memory to 12G in order to make testing on Vagrant
smoother.

Related-Bug: #1711376
Change-Id: Ib371c3513081a82271f18cde340e1cd9465a3da6