Remove repeated pip_install role execution
In the galera playbook, the pip_install role
is executed, then the galera_server role. The
galera_server role has a meta-dependency on
galera_client, which in turn has a
meta-dependency on pip_install. This results in
the pip_install role being run twice.

In the utility playbook, the galera_client role
is executed, then the pip_install role. The
galera_client role has a meta-dependency on
pip_install. This results in the pip_install
role being run twice.

This patch removes the duplicated executions of
the pip_install roles from these two playbooks.

Change-Id: Ie17b0154caa8851d99f0881e8596a7c39350bbe2