Add infra healthchecks
This adds a playbook, oriented for users, to automatically test
their installation after the setup infrastructure test.

These are connectivity tests to the infrastructure, and ensure
that containers would behave properly when running the openstack
playbooks.

These are based on the general assumption that ppl would use
a container management network.

Change-Id: Idb331dd6a72439c838708216500039f628b2760c