Update docs/tooling for Queens
In the upgrade tooling and the docs references we use
the series name in a few places. This updates them all
for development in the Queens cycle.

In the docs there are many references to Ocata manuals
which are left alone as the upstream Pike docs have not
been published yet.

Change-Id: Ice2169bd40f5712dd9ec9a9eb946fe0244f32294