scripts: scripts-library.sh: Fix location for log directory
The AIO job did not collect any logs from the host because the working
directory was changing due to pushd commands and the logs were placed
into the wrong location. We already have the root location in the
OSA_CLONE_DIR variable so lets use that and only try to detect it if
that variable is empty for whatever reason.

Change-Id: I17cc321ab640da4e251d21a996b537bda8cb3c1d