Fix LXC container start order
The lxc container configuration option to start containers in a
particular order is not working as expected. The ordering integer
was in reverse and if we do not set the "lxc.start.order" option on all
containers within a given host the LXC system start will start all
containers that do NOT have the option set first and then start the rest
of the containers using the start integer in numeric order. This change
inverts the start order integers we had set and defines a default start
order option at 100.

Change-Id: Ief6213d6c7c65f031d4c139db103e5c29e0db5ec
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>