playbooks: os-lxc-container-setup: Force kill a container on openSUSE if it takes too long to stop
[This is identical to I02299cb17477f63970e11ce3c8455cb12773a541]

It appears that lxc-1.1.5 is very slow when stopping a container. As a
result of which we pass the timeout (-t) option on openSUSE whilst this
is being investigated on the distribution level. This can also be
removed when lxc-2.X.X becomes the default option on openSUSE.

Link: https://bugzilla.opensuse.org/show_bug.cgi?id=1054609
Change-Id: Id9ba9a716a115df2ef3e7755f7ba43cc566683c7
(cherry picked from commit 76e1d6120ab8b3bea4430dae58ae0e18e978cc0c)