Correct zuul-cloner cloning of pinned SHA's
The zuul-cloner git clone process currently will always
clone the head of the particular branch rather than the
pinned SHA's.

This patch changes the process so that zuul cloner is
first used to get a base git repo in place using the
git cache, then the Ansible git module ensures that the
checkout to the correct SHA is done.

Change-Id: I9d0327f8f85cfa27b3113b3236b60c4c30262435