Update role SHA's for online migration fixes
Fixes have recently been implemented in order
to ensure that online migrations are properly
executed. This patch ensures that the fixes
are included in the next release.

Now that the migrations are executing correctly,
the missing nova_bin variable value has been
uncovered for stable.newton only and is also
added to this patch.

Change-Id: I1fac68013d0de5a4526ec8e63a0db2877ea6b0bb