scripts: scripts-library.sh: Fix tracepath logging
The 'a || b > foo' statement only logs stdout to the 'foo' file when
command 'b' is executed. If command 'a' is executed then output is
printed to the console so the file ends up being empty. As such we need
to group the OR statement so the redirection can be applied to the
entire group. Moreover, we hide the 'tracepath' warnings if we are using
an old 'tracepath' command so we can reduce the noise in the output.
(fix on top of Ife7f393f484212cf520e8f1cf337ef94df2a3ba4)

Change-Id: Id78c40e9d7471779eabcd1fcad1d433b7d364e58
(cherry picked from commit b3300e87f2933664eb809d56d15aba044029c078)