docs: add a decision help
There are a lot of different approaches to deploy OpenStack to your
environment. Namely Puppet, Chef, Helm, Juju, Fuel, TripleO, Kolla
and probably more. The designated OpenStack operators would benefit
from high-level decision help when and when not to chose openstack-
ansible. This does not imply that one project is "better" than the
others.
This change adds a few bullet points to the docs which might help
designated operators to make their decision, because switching that
decision when the OpenStack cloud is in production might not be that
easy.

Change-Id: I26e1b22010f56b1fb6eda7a02f9f31f79c3adc15