Limit group_vars scoping
Instead of defining almost all vars scoped to the 'all'group, we will
limit our vars scoping to the group namespace they are consumed in.

Vars that are used in 1 group are defined in the group they are used
in. This covers most service role vars.

Vars that are used in 2 groups (usually linking 2 roles communicating
with each other) are now using localhost's host_vars as a proxy. The
2 service groups will then reference hostvars['localhost']['var'].

Vars that are used in 3+ groups are left in the 'all' scoping.

Change-Id: I09cb22d51985d5f50e98030a455251cc2ea1a205