Fix cinder_service_region undefined
Due to the simplification of group vars [1], we've removed the
definition of cinder_service_region for magnum.

The magnum role has been simplified in a separate commit to remove the
dependency on this weak linki (scoping only inside its role),
still allowing it to be overriden by a deployer. (See dependant commit).

With that included, we only needed to wire the magnum_service_region
in group vars, so we wire the magnum_service_region inside the magnum
group to the standard "service_region" variable, which puts
everything at the same region, like all the other services are doing,
and also standardizes the configuration.

[1]: https://review.openstack.org/#/c/504804/18/group_vars/all/cinder.yml

Depends-On: Id25c2d344859adfd09108b759d55f11ae83f97a9
Change-Id: Ibda9da1feb319a4ceaae1d470d4e1ea53755c378