Update all SHAs for 15.1.11
This patch updates all the roles to the latest available stable
SHA's, copies the release notes from the updated roles into the
integrated repo, updates all the OpenStack Service SHA's, and
updates the appropriate python requirements pins.

On top of it, ceph-ansible 2.2.11 was pointing to some PPAs that
do not exist anymore. They were changed to include the version
number for nfs-ganesha. This bumps the version to a new tag that
includes the updated PPAs

Co-Authored-By: Sam Yaple <sam@yaple.net>

Depends-On: Id5245ca94ceacb488bf55c4dba0fbc296f0dd304
Change-Id: Ibbba66213782eadae9064fa5d790a181ddbf8d46