Include the Ansible version variable when installing ara
When installing ARA the ansible version we're using needs to be provided
too. Without specifying the ansible version ARA has the ability to
install whatever it wants which could override our already set version
of ansible. If this happens we'll see very unexpected results in test.

Example: ara==0.14.0 will attempt to pull in ansible 2.4 when building
         a pike AIO which results in a failed build.
         https://github.com/openstack/openstack-ansible/blob/stable/pike/scripts/gate-check-commit.sh#L117

By providing the ANSIBLE_PACKAGE variable along with ARA we'll be able
to guard against issues caused by unbound constraints.

Change-Id: I2443242f285381037a0351ceb4e6d997271dbd4b
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>