Adapt to openstack_hosts/pip_install changes
The commit [1] introduced a change in the openstack_hosts role,
allowing it to be aware of group_vars/host_vars.

We therefore don't need the group matching in the variable
``openstack_host_specific_kernel_modules``.

On top of that, the commit [2] allowed repo management for all
hosts where openstack_hosts is applied, and commit [3] removed
UCA from pip install, which applies to all the hosts,
whether they are containers or not.
We should therefore add UCA to containers back.

[1]: https://review.openstack.org/#/c/515397
[2]: https://review.openstack.org/#/c/515399
[3]: https://review.openstack.org/#/c/515400/

Change-Id: Ic959a42f7c6fb33b80e646b6689fc2d54f4f9ac8