[Docs] Fix links to inventory docs
The documentation on generating inventory was referenced as
'generate-inventory' in a table of contents index but the actual file
was named 'inventory' so it was never listed. The file has been renamed.

Also, a page in the developer docs included duplicated information from
the inventory docs. That page has been deleted and the link to it has
been replaced with one to the more complete inventory docs.

Change-Id: I6e1cda1daa1931c925b93872565994109216e7ea