Update aio job to use become and fix environment
The in-tree job definition did not pass the legacy ZUUL environment
into sudo or dump the ZUUL_ prefixed vars into zuul.env, so the
get-ansible-role-requirements script did not detect that it should
use Zuul provided sources when available.

This fixes the job definition in several ways:
1) The job definition no longer calls gate-check-commit using the
   sudo command. Instead become/become_user are placed on the task
   so ansible will handle the sudoing and we just call
   gate-check-commit directly now.

2) The zuul legacy environment is removed from the job environment.
   Only ZUUL_PROJECT is used in the integrated repo[1], so that is
   the only var we define now, and we do not need anything special to
   pass it into gate-check-commit since it is no longer running in
   a sudo wrapper.

As a result of these changes, the zuul.env loading code is removed
from gate-check-commit.

[1] https://github.com/openstack/openstack-ansible/blob/c3ce03873798cd4e71b3fdd4ee9a2eae4bf59d7c/tests/get-ansible-role-requirements.yml#L57

Change-Id: Id0fb8502029e3f996d2557bc09c3314cc884bba8