Remove legacy zuul environment sourcing
Only ZUUL_PROJECT is used in the integrated repo[1], so that is
the only var we define now, and we do not need anything special to
pass it into gate-check-commit since it is no longer running in
a sudo wrapper.

As a result of these changes, the zuul.env loading code is removed
from gate-check-commit.

[1] https://github.com/openstack/openstack-ansible/blob/c3ce03873798cd4e71b3fdd4ee9a2eae4bf59d7c/tests/get-ansible-role-requirements.yml#L57

Backport note:
This is a partial backport of https://review.openstack.org/516383.
The change to zuul.d/playbooks/run.yml is not necessary to include
because it was introduced in I8d90c3ea149ca356cea92e439e04c539ee6dd719

Change-Id: Id0fb8502029e3f996d2557bc09c3314cc884bba8
(cherry picked from commit 0405ec9f76203bce9e6e12e32ff516fe7b7df83e)