Fix cinder_backup for horizon dashboard and heat
We specify the "horizon_enable_cinder_backup" and the
"heat_cinder_backups_enabled" settings to be the value of
"cinder_service_backup_program_enabled" if defined. However that
variable is never exposed to the horizon or heat group_vars, so it will
always be undefined, and the feature will never be enabled for heat or
horizon.

We should set these to the variable on localhost, which is how we specify
it for tempest and cinder, since that variable exists as a host_var it
will always exist.

Change-Id: Ibf2a32e03733185fa4bfe9274c65050e12317e6b