Properly detect RBD in use on Cinder backends
The Cinder RBD in use host_var was not being set correctly when RBD
backends are defined in extra vars files. Instead we should use a
set_fact as implemented in Pike and master, instead of the current
broken implementation in Ocata.

A similar set fact is also added to the cinder playbook. This set
fact does not exist in Pike and master since cinder's glance API
version setting is no longer a ternary 1/2 switch, the setting is
removed since Cinder uses Glance v2 by default in Pike+. We add the
set fact here since Ocata still has a ternary switch to determine
which glance API version to use.

Change-Id: I29700603fe0ad953cde6239e4cace77c29013331