Fix port config in haproxy to avoid conflict
Opendaylight and repo services are using the same port in the frontend. That
creates connectivity problems. The patch changes the port for opendaylight and
a new variable (haproxy_backend_port) is added so that we can still use the
service running in the old port (8181) relying on haproxy to do a redirection

Change-Id: Ife8c424f26685a6957a9bb7cdccc2ce0523fd3fb