Add dynamic table for our tested Scenarios
This change adds in a dynamic table to our documentation so that we're
able to generate a table of all available scenarios and mark where
services are tested within. To make the dynamic scenario generation
easier the AIO scenario conf setup has been moved into it's own variable
file.

The file doc/source/contributor/scenario-table-gen.html has been touched
allowing sphinx to import and generate the required table as needed.

Change-Id: Ief5d180491d3a330890d6bfa06d43f565dd26103
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>