Change the galera health check for better cluster health
The current galera cluster health check simply logs into a cluster node
but does not check if the node is sync'd. This can lead to an issue
where a node is placed back into the pool before it is ready. If this
happens it can lead to a broken OpenStack environment until the wsrep
received queue is processed which is especially true if the node out of
sync happens to be the primary.

Change-Id: I49e371a2743618a0b5544a23e892aa28bb8567eb
Closes-Bug: #1665667
Depends-On: Ie1b3b9724dd33de1d90634166e585ecceb1f4c96
Depends-On: I81c924464aa4b19c2a62f37b5bf26c3c0453786a
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>