Fix crontab errors on CentOS/SUSE
CentOS comes with a default empty crontab file that is fine for
most deployers. It also lacks the `--report` option for `run-parts`
that Ubuntu has.

SUSE doesn't have run-parts at all.

This causes the server administrator to receive emails with errors
from cron on a regular basis. This patch limits the crontab
deployment to Ubuntu only and skips CentOS/SUSE hosts.

Closes-Bug: 1732762
Change-Id: Iba714e1e20e1a27b1a5bb58a714dee3916faba8f