Add ceph_mons to ceph example
The ceph documentation backported to Pike does not include ceph_mons
in the vars example, because the var does not need to be defined in
master.

In Pike however, it does need to be defined, and ceph_client will not
run unless it is defined by the user. Once
Ida44228c2d2919990bac46936317b6093d3ff115 has merged, this commit
can be reverted.

Change-Id: Ia6218dbc90af3e379af28bff5cc957dd3feb81d0