Allow pip_install to run in repo-server playbook
In order to switch from having a repo server which
has all the python packages on it to a caching server
which *optionally* has packages on it, we make use of
a combination of pypiserver and nginx in a reverse
proxy configuration.

This patch implements the changes to ensure that we
are able to do pip installs in the repo_server role
and to transport the pypiserver logs via rsyslog to
the central log server.

The pip_offline_install var is also set in group_vars
as it's used in group_vars, but is only set in the
pip_install role.

Change-Id: Id20a43fed833d53ca0f147f517deafba6587352d
Implements: blueprint python-build-install-simplification
Signed-off-by: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>