Do not implement reload for novnc/spice service
The nova-novncproxy and nova-spicehtml5proxy services do
not support using a HUP signal to reload, so they need to
be restarted instead.

This patch ensures that these services are restarted, and
also fixes a bug that hasn't been noticed before where if
the Placement API and other nova services are implemented
on the same host/container then the other services would
not be restarted. It would play out like this:

- The first task asks to reload all nova services execute
  against all nova hosts, except those running the placement
  api. This results in only the compute hosts getting the
  reload.

- The second task only asks the nova placement API to
  restart. This leaves out all the other nova services
  which are on the same host.

Change-Id: Id4b9caa1cdfd0f53023548c4ede5d497d64c873d
Needed-By: I36e2200e2ca43d87208d2c4edb17b98654bb121e