Allow neutron_provider_networks to be user defined
The neutron_provider_networks variable in the os_neutron role is
generated however there are times when a deployer might want to set this
option explicitly. This change gives us the ability to define
neutron_provider_networks up-front in something like a group var. If the
value is undefined it will simply carry on generating the data.

Change-Id: I232bee6f20a0221e1519c2f22e7f8ec19d5facfd
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>