Allow disabling container restarts
Allow a deployer to disable container restarts by setting
``lxc_container_allow_restarts`` to ``false``.

The tasks in common-tasks/os-lxc-container-setup.yml change some
container options that require a restart to be effective. If these
change during an upgrade this can cause uncoordinated LXC container
restarts of galera or rabbitmq containers.

This is a complement to the same option already present in the
lxc_container_create role.

Change-Id: Ib9b9daf816b0f62522997d2b0e66b6a5acd9a645