Use correct name for networking-odl git group
Currently the networking-odl repo will always be cloned,
and the package always built due to the incorrect naming
of the project group. The group used thus falls back to
the default of 'all', resulting in this behaviour.

This patch corrects the name, but also moves the git repo
variable details into the openstack_services file along
with the other neutron drivers/plugins.

The spacing between the variables is also corrected so that
all related variables have a single line break between them,
but all unrelated items have a double line break between
them.

Change-Id: I8786405d6174e69bef5d85b8ff473d2d262d068b