Update test tooling for manual testing
In https://review.openstack.org/520065 the lint tests
were updated to make use of the tests repo for all
linting. This works nicely within OpenStack-CI, but
fails when trying to do it by hand due to missing
packages on the hosts. It also requires the tester to
install some things beforehand.

To align this repo with the way the role tests run,
we implement the same mechanism as used in the roles
to prep the host using the run_tests.sh script and the
appropriate bindep entries.

Change-Id: Ia1c0a7374d8027d0f0772f393b462ee807ea2090